Move LuceneQueryComponent from core to lucene.
